data!: Set empty string values for certified products to null
This is done to "normalize" all the existing data. Business logic in the
normalizer classes of code will take care of setting all future values
to null if they are empty strings, but this is done to adjust current
data and avoid activity being created when the business code sets a
value to null when it used to be an empty string.

[#OCD-4282]