feat!: Change certification status calculation for many views
Lots of our views calculate the current certification status of a
listing. Add a function to calculate the status based on "now". It used
to be based on the most recent status event, but that won't work if the
most recent status event has a future date. The function uses the max
status event that has a date on or before "now". Certification status
events with dates after "now" will not be considered.

[#OCD-4452]