data!: Remove 'statuses' field from old developer activity
There used to be a field called 'statuses' in a DeveloperDTO object that
really should have had a more descriptive name (it was actually a
summary of the counts of listings with each status under each
developer). That field was removed from use some time ago. In OCD-4554,
we are adding a new field and we want to call it 'statuses'. It is an
array of stuff rather than a single object like the old field and the
API parsing of developer activity failed. Since we don't need these old
statuses values for any historical purpose and we never really used them
we decided to get rid of them in the old activity.

[#OCD-4554]