data!: Backfill notification timestamp for subscription observations
If a subscription is not deleted, then the "last modified date" of each
observation correlates to the time that the observation was marked as
deleted (which is the same time the user was notified). We can use this
information to backfill the notified_at column for all existing
observations.

Currently there are no "deleted" subscriptions. It's likely people don't
care enough to unsubscribe.

[#OCD-4483]