data!: Clean up multiple spaces in user-entered data
Some fields that users can enter as free-text were found to have
multiple spaces internally. We were already trimming the fields on the
ends, but we also want to clean up these additional spaces in the middle
of the data. We do not believe the spaces have meaning and they have
never been visible in the UI because HTML displays infinite contiguous
spaces as one single space.

[#OCD-4690]