test[aqa]: Add output at end of load script
The AQA tests started failing because (I think) the last line printed
out from the pg_restore command had the word "error" in it, though it
was not an error message. I think Bamboo parsed that and detected it as
an error and the AQA tests would stop on the step of loading the prod
db.

[#OCD-4485]